fix: RQ kernel Int/Long — all 7 quant types pass
RQ decode kernel needed .int() on block_table/seq_lens (same fix as TQ).

All quant types verified:
  tq3:  cos=0.925  tq4:  cos=0.974
  tq3w: cos=0.985  tq4w: cos=0.997  ← WHT best
  rq2:  cos=0.802  rq3:  cos=0.924  rq4: cos=0.970

+++ vllm/v1/attention/backends/multiquant_attn.py
@@ -699,8 +699,8 @@ class MultiQuantImpl:
                 rq_kernel.rq_fused_decode_attention(
                     q_rot_3d, q_proj_3d, kv_cache,
                     Pi, S, centroids,
-                    attn_metadata.block_table[:num_decode],
-                    attn_metadata.seq_lens[:num_decode],
+                    attn_metadata.block_table[:num_decode].int(),
+                    attn_metadata.seq_lens[:num_decode].int(),
                     rb["rq_out"],
                     D, self._mse_bits, centroids.shape[0], self.scale,
                     s_block, s_kv, s_slot, s_head,
